pyproject.toml
[project]
name = "oregon-notification-service"
version = "0.1.0"
description = "notifications"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiokafka>=0.13.0",
    "alembic>=1.13.0",
    "asyncpg>=0.31.0",
    "confluent-kafka>=2.14.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.0",
    "sse-starlette>=3.3.4",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.28.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
    "ty>=0.0.21",
]



[tool.ruff]
target-version = "py314"
line-length = 100
include = ["src/**/*.py", "tests/**/*.py"]
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.pyi",
    "migrations",
    "node_modules",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "SIM",
    "RET",
    "ARG",
    "PTH",
    "RUF",
    "ANN",
    "ERA",
    "G",
]
ignore = [
    "E501",
    "ARG001",
]
fixable = ["ALL"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v"

[tool.pytest_asyncio]
asyncio_mode = "auto"
default_fixture_loop_scope = "function"

[tool.ty.src]
include = ["src", "tests"]
